feat(v1.0.2): D3 pitch compensation, envelope filter, MONOBASS debounce, UI improvements
- Add pitch-based volume compensation for D3 606/FM voices (boost at low pitch)
- Cap 606 max pitch at 4000 Hz, add band-pass tracking to tame harshness
- Envelope filter with 3-zone curve (deadband/crossfade/full) for D1 chroma/MONOBASS
- Link attack knob to envelope filter tau for breathing filter sweeps
- MONOBASS integrator debounce (5-sample) replacing timer-based approach
- MONOBASS grid shows once per power cycle, volume boost, env filter support
- Dual-mode master filter: LP below noon, OFF at noon, HP above noon
- Save requires double-press confirmation within 2s window
- Chroma toggle: 120ms animation, background layer, 12x12 squares
- Parameter overlay centered in scope window with 2px solid text outlines
- D3 voice trims rebalanced (606=6.0, FM=3.5, perc=0.15)
- D1 attack range 0.5-250ms, LP floor 2500 Hz, HP max 350 Hz, D3 LPF floor 3000 Hz
- Chroma entry snap reduced to 50%

+++ ext_sync.h
@@ -255,7 +255,7 @@ void externalClockISR() {
     // The (int32_t) cast makes the subtraction signed so >> 1 rounds toward zero.
     extIntervalEMA = (ema == 0)
         ? interval
-        : ema + (((int32_t)(interval - ema)) >> 1);
+        : ema + (((int32_t)interval - (int32_t)ema) >> 1);
 
     // Slow EMA for subdivision placement — doesn't chase jitter like the fast EMA.
     // At 80 BPM / 2 PPQN, 5 ms of pulse jitter shifts synthInterval by only 0.625 ms.
